# specs/001-password-manager/contracts/openapi.yaml
openapi: 3.1.0
info:
  title: Password Manager API
  version: 1.0.0
  description: >
    Zero-knowledge password manager. The server stores and authorizes ciphertext it cannot read.
    Every field with `format: envelope` is client-encrypted per contracts/crypto-envelope.md.
    See contracts/README.md for the error model, authorization matrix, and known limitations.
servers:
  - url: /api/v1
security:
  - sessionCookie: []
...
paths:
...
  /vaults/{vaultId}/secrets:
    parameters: [{ $ref: '#/components/parameters/VaultId' }]
    get:
      tags: [secrets]
      summary: List all secrets in the vault as ciphertext
      description: >
        Returns the full set for client-side indexing and search (research.md §4). There is no
        server-side search endpoint because the server cannot read titles or values.
      responses:
        '200':
          content:
            application/json:
              schema: { type: array, items: { $ref: '#/components/schemas/Secret' } }
    post:
      tags: [secrets]
      summary: Create a secret (Owner or Editor)
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/SecretWrite' }
      responses:
        '201':
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Secret' }
        '403': { description: ROLE_INSUFFICIENT — a Viewer may not write (SC-012) }
  /vaults/{vaultId}/secrets/{secretId}:
    parameters:
      - { $ref: '#/components/parameters/VaultId' }
      - { name: secretId, in: path, required: true, schema: { type: string, format: uuid } }
    put:
      tags: [secrets]
      summary: Update a secret with optimistic concurrency
      requestBody:
        required: true
        content:
          application/json:
            schema:
              allOf:
                - { $ref: '#/components/schemas/SecretWrite' }
                - type: object
                  required: [revision]
                  properties:
                    revision: { type: integer }
      responses:
        '200': { description: Updated }
        '409':
          description: >
            REVISION_CONFLICT — another member saved first. The current row is returned so the client
            can present a merge rather than silently discarding either edit.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Secret' }
    delete:
      tags: [secrets]
      summary: Soft-delete a secret (recoverable for 30 days, FR-046)
      responses:
        '204': { description: Deleted }
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: pm_session
  parameters:
    VaultId:
      name: vaultId
      in: path
      required: true
      schema: { type: string, format: uuid }
...
  schemas:
...
    SecretWrite:
      type: object
      required: [title, templateVersionId, fieldValues]
      properties:
        title: { type: string, format: envelope }
        templateVersionId: { type: string, format: uuid }
        folderId: { type: string, format: uuid, nullable: true }
        tagIds: { type: array, items: { type: string, format: uuid } }
        fieldValues:
          type: object
          description: Map of fieldId to value. Sensitive fields are envelopes; others are plain.
          additionalProperties: true
    Secret:
      allOf:
        - { $ref: '#/components/schemas/SecretWrite' }
        - type: object
          properties:
            id: { type: string, format: uuid }
            vaultId: { type: string, format: uuid }
            keyVersion: { type: integer }
            revision: { type: integer }
            updatedAt: { type: string, format: date-time }
            deletedAt: { type: string, format: date-time, nullable: true }
...